requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.4
colorama==0.4.4
commonmark==0.9.1
defusedxml==0.7.1
idna==3.2
onelogin==2.0.2
Pygments==2.10.0
python-dateutil==2.8.2
python-dotenv==0.19.0
requests==2.26.0
rich==10.7.0
six==1.16.0
urllib3==1.26.6